Changed ids for i32

+++ src-tauri/src/controllers/question_controller/delete.rs
@@ -7,7 +7,7 @@ use crate::models::{
 };
 
 #[tauri::command]
-pub fn delete_question(id: u64, db: tauri::State<Database>) -> BackendResponse<String> {
+pub fn delete_question(id: i32, db: tauri::State<Database>) -> BackendResponse<String> {
     let rw = match db.rw_transaction() {
         Ok(rw_t) => rw_t,
         Err(err) => {
